refactor: rename components folder to demo modules
