+++ demos/demo-modules/another/component.ts
