+++ demos/demo-modules/kitchen-sink/component.ts
