+++ demos/demo-modules/kitchen-sink/module.ts
